[#936] Move ICCProfileInvalidError/ICCProfile into profile.py (item 10)
DisplayCAL/icc_profile/__init__.py is now a pure re-export surface: every
name it exposes comes from a submodule import, none of it locally defined.
ICCProfileInvalidError and ICCProfile (with the module's _ICCPROFILE_CACHE)
move to icc_profile/profile.py verbatim.

Also fixes a regression from item 9: create_synthetic_hdr_clut_profile's
extraction dropped the TYPE_CHECKING-only `import threading`, but
ICCProfile.apply_black() (still in __init__.py at the time) also uses
`threading.Event` as a type hint, so profile.py needs it back.

tests/test_ICCProfile.py's ICCPROFILES_HOME/ICCPROFILES monkeypatching now
targets DisplayCAL.icc_profile.profile directly (same module-attribute
gotcha as item 8's test_icc_profile_hdr_filtering.py fix), since ICCProfile
now reads those names from profile.py's globals, not icc_profile's.
